Reenabled -Wconversion and fixed some compiler warnings

+++ src/editor/combobox.cpp
@@ -145,7 +145,7 @@ Combobox::get_box_offset()
 bool
 Combobox::set_selected_item(int id)
 {
-  for(size_t i = 0; i < item_list.size(); ++i)
+  for(int i = 0; i < static_cast<int>(item_list.size()); ++i)
   {
     if (item_list[i].id == id)
     {
